fix bug - if url length > 22, URI would be wrong
If you run "mifareclassic_formatndef" or "mifareclassic_updatendef" examples with an url longer than 22 character, before this commit your program would enter the if loop between line 638 and line 642, which cause NDEF reader not even able to recognize the formatted card. Also indexing in line 652 and 653 was not correct either.